feat(Lab04): implement remove() function
The remove() function first validates whether a record description
was input by the user before checking whether any line matching the
description was present in the .finances file. If there was one or
more matches, 'grep -v' is run on the file with the user's input
description pattern to produce a string which contains all lines of
the .finances file, not including those that match the description.
The .finances file is then overwritten with this string. In either
case, a status message is output to the console.